requirements.txt
# Azure SDK dependencies
azure-cognitiveservices-speech==1.38.0
azure-ai-textanalytics==5.3.0
azure-storage-blob==12.19.0
azure-functions==1.18.0

# Audio processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Web API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Whisper transcription
openai-whisper==20231117
openai==1.10.0

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0

# Type checking and linting
mypy==1.10.0
black==24.4.2
flake8==7.0.0
